feat: extract routeAssetId matchers
