+++ src/hooks/useRouteAssetId/useRouteAssetId.ts
@@ -1,45 +1,33 @@
 import { AssetId } from '@shapeshiftoss/caip'
+import { getFoxPageRouteAssetId } from 'plugins/foxPage/utils/getFoxPageRouteAssetId'
 import { useEffect, useState } from 'react'
 import { matchPath, useLocation } from 'react-router'
 
-const FoxRoutePartToAssetId: Record<string, string> = {
-  fox: 'eip155:1/erc20:0xc770eefad204b5180df6a14ee197d99d808ee52d',
-  foxy: 'eip155:1/erc20:0xdc49108ce5c57bc3408c3a5e95f3d864ec386ed3',
-}
+const getRouteAssetId = (pathname: string) => {
+  // Extract the chainId and assetSubId parts from an /assets route, see src/Routes/RoutesCommon.tsx
+  const assetIdPathMatch = matchPath<{ chainId: string; assetSubId: string }>(pathname, {
+    path: '/assets/:chainId/:assetSubId',
+  })
 
-const FoxPageDefaultAsset = 'fox'
+  if (assetIdPathMatch?.params) {
+    const { chainId, assetSubId } = assetIdPathMatch.params
+
+    // Reconstitutes the assetId from valid matched params
+    const assetId = `${chainId}/${assetSubId}`
+    return assetId
+  }
+}
 
 export const useRouteAssetId = () => {
   const location = useLocation()
   const [assetId, setAssetId] = useState<AssetId>('')
 
   useEffect(() => {
-    // Extract the chainId and assetSubId parts from an /assets route, see src/Routes/RoutesCommon.tsx
-    const assetIdPathMatch = matchPath<{ chainId: string; assetSubId: string }>(location.pathname, {
-      path: '/assets/:chainId/:assetSubId',
-    })
+    const routeAssetId = getRouteAssetId(location.pathname)
+    const foxPageRouteAssetId = getFoxPageRouteAssetId(location.pathname)
 
-    // Extract the foxAsset part of a /fox/:foxAsset route, can be 'fox', 'foxy' or undefined
-    const foxPageAssetIdPathMatch = matchPath<{ foxAsset?: string }>(location.pathname, {
-      path: '/fox/:foxAsset?',
-    })
-
-    if (foxPageAssetIdPathMatch) {
-      const foxAsset = foxPageAssetIdPathMatch?.params?.foxAsset ?? FoxPageDefaultAsset
-
-      if (FoxRoutePartToAssetId[foxAsset]) {
-        setAssetId(FoxRoutePartToAssetId[foxAsset])
-      }
-      return
-    }
-
-    if (assetIdPathMatch?.params) {
-      const { chainId, assetSubId } = assetIdPathMatch.params
-
-      // Reconstitutes the assetId from valid matched params
-      const assetId = `${chainId}/${assetSubId}`
-      setAssetId(assetId)
-      return
+    if (routeAssetId || foxPageRouteAssetId) {
+      setAssetId(routeAssetId ?? foxPageRouteAssetId ?? '')
     }
   }, [location.pathname])
