+++ src/plugins/foxPage/utils/getFoxPageRouteAssetId.ts
@@ -1,0 +1,22 @@
+import { matchPath } from 'react-router'
+
+const FoxRoutePartToAssetId: Record<string, string> = {
+  fox: 'eip155:1/erc20:0xc770eefad204b5180df6a14ee197d99d808ee52d',
+  foxy: 'eip155:1/erc20:0xdc49108ce5c57bc3408c3a5e95f3d864ec386ed3',
+}
+
+const FoxPageDefaultAsset = 'fox'
+
+export const getFoxPageRouteAssetId = (pathname: string) => {
+  const foxPageAssetIdPathMatch = matchPath<{ foxAsset?: string }>(pathname, {
+    path: '/fox/:foxAsset?',
+  })
+
+  if (foxPageAssetIdPathMatch) {
+    const foxAsset = foxPageAssetIdPathMatch?.params?.foxAsset ?? FoxPageDefaultAsset
+
+    if (FoxRoutePartToAssetId[foxAsset]) {
+      return FoxRoutePartToAssetId[foxAsset]
+    }
+  }
+}
